refactor(react-signals): fix protected methods
