+++ packages/ts/react-signals/test/FullStackSignal.spec.tsx
@@ -5,7 +5,8 @@ import { ConnectClient, type Subscription } from '@vaadin/hilla-frontend';
 import { nanoid } from 'nanoid';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
-import { DependencyTrackingSignal, type SnapshotStateEvent, type StateEvent } from '../src/FullStackSignal.js';
+import type { SnapshotStateEvent, StateEvent } from '../src/events.js';
+import { DependencyTrackingSignal } from '../src/FullStackSignal.js';
 import { computed, NumberSignal } from '../src/index.js';
 import { nextFrame } from './utils.js';
